fixing missing fields in meta form
